fix(react-color-picker): move CSS vars to constants
