+++ change/@fluentui-react-color-picker-8c7d12ab-91de-4d96-a0bc-0f23c466b401.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "fix: Move shared color picker CSS variables from style modules to constants files.",
+  "packageName": "@fluentui/react-color-picker",
+  "email": "dmytrokirpa@microsoft.com",
+  "dependentChangeType": "patch"
+}
