+++ packages/react-components/react-color-picker/library/src/components/AlphaSlider/AlphaSlider.constants.ts
@@ -1,0 +1,6 @@
+export const alphaSliderCSSVars = {
+  sliderDirectionVar: `--fui-AlphaSlider--direction`,
+  sliderProgressVar: `--fui-AlphaSlider--progress`,
+  thumbColorVar: `--fui-AlphaSlider__thumb--color`,
+  railColorVar: `--fui-AlphaSlider__rail--color`,
+};
