+++ packages/react-components/react-color-picker/library/src/components/AlphaSlider/useAlphaSliderStyles.styles.ts
@@ -5,6 +5,7 @@ import { tokens } from '@fluentui/react-theme';
 import type { SlotClassNames } from '@fluentui/react-utilities';
 import { useColorSliderStyles_unstable } from '../ColorSlider/useColorSliderStyles.styles';
 import type { AlphaSliderSlots, AlphaSliderState } from './AlphaSlider.types';
+import { alphaSliderCSSVars } from './AlphaSlider.constants';
 
 const TRANSPARENT_IMAGE_URL = 'https://fabricweb.azureedge.net/fabric-website/assets/images/transparent-pattern.png';
 
@@ -13,13 +14,6 @@ export const alphaSliderClassNames: SlotClassNames<AlphaSliderSlots> = {
   rail: 'fui-AlphaSlider__rail',
   thumb: 'fui-AlphaSlider__thumb',
   input: 'fui-AlphaSlider__input',
-};
-
-export const alphaSliderCSSVars = {
-  sliderDirectionVar: `--fui-AlphaSlider--direction`,
-  sliderProgressVar: `--fui-AlphaSlider--progress`,
-  thumbColorVar: `--fui-AlphaSlider__thumb--color`,
-  railColorVar: `--fui-AlphaSlider__rail--color`,
 };
 
 /**
