+++ packages/react-components/react-color-picker/library/src/components/ColorArea/ColorArea.constants.ts
@@ -1,0 +1,6 @@
+export const colorAreaCSSVars = {
+  areaXProgressVar: `--fui-AreaX--progress`,
+  areaYProgressVar: `--fui-AreaY--progress`,
+  thumbColorVar: `--fui-Area__thumb--color`,
+  mainColorVar: `--fui-Area--main-color`,
+};
